Start working on overhauling the top profile section in settings
