+++ _locales/_untranslated_en.json
@@ -13,5 +13,8 @@
   },
   "import_qr_error": {
     "message": "QR could not be decoded"
+  },
+  "pref_edit_profile": {
+    "message": "Edit profile"
   }
 }
