+++ scss/dialogs/_settings.scss
@@ -42,42 +42,83 @@
   cursor: pointer;
 }
 
-.profile-image-selector {
-  float: right;
-  width: 150px;
+
+.profile-image-username {
   display: flex;
-  flex-direction: column;
-  align-items: center;
-  & > img,
-  & > span {
+  flex-direction: row;
+  &.center {
+    align-items: center;
+    flex-direction: column;
+  }
+  .profile-image-selector {
     width: 100px;
-    height: 100px;
-    border-radius: 100%;
-    margin: 5px;
-    user-select: none;
-    pointer-events: none;
-  }
-  & > span {
-    line-height: 100px;
-    font-size: 64px;
-    color: white;
-    text-align: center;
-  }
-  & > div {
-    // the button container
-    & > button {
-      &:first-child {
-        border-top-right-radius: 0;
-        border-bottom-right-radius: 0;
-        border-right: none;
+    display: flex;
+    flex-direction: column;
+
+
+    & > img,
+    & > span {
+      width: 100px;
+      height: 100px;
+      border-radius: 100%;
+      user-select: none;
+      pointer-events: none;
+    }
+    & > span {
+      line-height: 100px;
+      font-size: 64px;
+      color: white;
+      text-align: center;
+    }
+    
+    .part-circle {
+      top: -20px;
+      position: relative;
+      height: 40px;
+      overflow: hidden;
+      visibility: hidden;
+      margin-bottom: -40px;
+      &.visible {
+        visibility: visible;
       }
-      &:last-child {
-        border-top-left-radius: 0;
-        border-bottom-left-radius: 0;
+      .circle {
+        position: relative;
+        top: -80px;
+        width: 100px;
+        height: 100px;
+        overflow: hidden;
+        border-bottom-left-radius: 30px;
+        border-bottom-right-radius: 100%;
+        border-radius: 100%;
+        /* border-bottom-right-radius: 1000px; */
+        background-color: #d3d3d3ab;
+        .camera-icon {
+          &::before {
+            content: 'x'
+          }
+        }
       }
+    }
+    
+  }
+  .profile-displayname-addr {
+    display: flex;
+    margin-left: 20px;
+    flex-direction: column;
+    justify-content: center;
+
+    .displayname {
+      font-size: large;
+      font-weight: 200;
+    }
+
+    .addr {
+      color: #969696;
+      margin-top: 5px;
     }
   }
 }
+
 
 .theme-meta {
   background-color: #80808082;
