+++ src/main/deltachat/burnerAccounts.ts
@@ -1,4 +1,3 @@
-import fetch from 'node-fetch'
 // import { getLogger } from '../../shared/logger'
 // const log = getLogger('main/deltachat/burner')
